requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
python-dotenv
requests==2.31.0